fix: using react hooks to watch the element updates
