+++ packages/core/src/runtime/Content.tsx
@@ -1,6 +1,7 @@
 import { ReactNode, Suspense } from 'react';
 import { matchRoutes, useLocation } from 'react-router-dom';
 import { normalizeRoutePath } from './utils';
+import { useViewTransition } from './hooks';
 
 const { routes } = process.env.__SSR__
   ? (require('virtual-routes-ssr') as typeof import('virtual-routes-ssr'))
@@ -13,9 +14,14 @@ export const Content = ({ fallback = <></> }: { fallback?: ReactNode }) => {
     return <div></div>;
   }
   const routesElement = matched[0].route.element;
+  /**
+   * useLayoutEffect to flush elements animation
+   */
+  const element = useViewTransition(routesElement);
+
   // React 17 Suspense SSR is not supported
   if (!process.env.__IS_REACT_18__ && process.env.__SSR__) {
-    return routesElement;
+    return element;
   }
-  return <Suspense fallback={fallback}>{routesElement}</Suspense>;
+  return <Suspense fallback={fallback}>{element}</Suspense>;
 };
