+++ packages/core/src/theme-default/components/Link/index.tsx
@@ -4,7 +4,6 @@ import nprogress from 'nprogress';
 import { routes } from 'virtual-routes';
 import styles from './index.module.scss';
 import { normalizeHref, normalizeRoutePath, withBase } from '@/runtime';
-import { useNavigateWithTransition } from '@/theme-default/logic';
 
 export interface LinkProps {
   href?: string;
@@ -23,7 +22,7 @@ export function Link(props: LinkProps) {
   const target = isExternal ? '_blank' : '';
   const rel = isExternal ? 'noopener noreferrer' : undefined;
   const withBaseUrl = isExternal ? href : withBase(normalizeHref(href));
-  const navigate = useNavigateWithTransition();
+  const navigate = useNavigate();
   const handleNavigate = async (
     e: React.MouseEvent<HTMLAnchorElement, MouseEvent>,
   ) => {
