+++ packages/core/src/theme-default/logic/index.ts
@@ -4,5 +4,4 @@ export { useSidebarData } from './useSidebarData';
 export { useLocaleSiteData } from './useLocaleSiteData';
 export { setup, bindingAsideScroll, scrollToTarget } from './sideEffects';
 export { usePathUtils } from './usePathUtils';
-export { useNavigateWithTransition } from './useNavigateWithTransition';
 export * from './utils';
